// src/types/feature.ts
export interface Character {
    name: string;
    height: string;
    mass: string;
    hair_color: string;
    skin_color: string;
    eye_color: string;
    birth_year: string;
    gender: string;
    homeworld: string; 
    films: string[];
    species: string[]; 
    vehicles: string[];
    starships: Starship[];
    created: string; 
    edited: string;  
    url: string;
}

export interface Planet {
    name: string;
    rotation_period: string; 
    orbital_period: string;   
    diameter: string;         
    climate: string;
    gravity: string;
    terrain: string;
    surface_water: string;    
    population: string;       
    residents: string[];       
    films: string[];            
    created: string;           
    edited: string;           
    url: string;
}

interface Film {
    url: string;
}

export interface Starship {
    name: string;
    model: string;
    manufacturer: string;
    cost_in_credits: string; 
    length: string;          
    max_atmosphering_speed: string; 
    crew: string;           
    passengers: string;     
    cargo_capacity: string;
    consumables: string;
    hyperdrive_rating: string; 
    MGLT: string;            
    starship_class: string;
    pilots: string[];        
    films: Film[];          
    created: string;         
    edited: string;         
    url: string;
}

export type featureData = Character | Starship | Planet;